+++ src/lib.rs
@@ -84,6 +84,8 @@ use winapi::shared::{
 };
 
 mod raw;
+#[cfg(windows)]
+pub mod sendqueue;
 #[cfg(feature = "capture-stream")]
 pub mod stream;
 
@@ -115,6 +117,8 @@ pub enum Error {
     InvalidRawFd,
     /// Errno error
     ErrnoError(errno::Errno),
+    /// Buffer size overflows capacity
+    BufferOverflow,
 }
 
 impl Error {
@@ -142,6 +146,7 @@ impl fmt::Display for Error {
             #[cfg(not(windows))]
             InvalidRawFd => write!(f, "invalid raw file descriptor provided"),
             ErrnoError(ref e) => write!(f, "libpcap os errno: {}", e),
+            BufferOverflow => write!(f, "buffer size too large"),
         }
     }
 }
@@ -162,6 +167,7 @@ impl std::error::Error for Error {
             #[cfg(not(windows))]
             InvalidRawFd => "invalid raw file descriptor provided",
             ErrnoError(..) => "internal error, providing errno",
+            BufferOverflow => "buffer size too large",
         }
     }